requirements.txt
python-dotenv>=1.0.0
langchain-google-genai>=0.0.6
langchain>=0.1.0
google-generativeai>=0.3.0
numpy>=1.24.0
pandas>=2.0.0
langchain>=0.1.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
beautifulsoup4
requests~=2.32.3
yfinance~=0.2.61
streamlit~=1.45.1
pyttsx3~=2.98
SpeechRecognition
pyaudio
whisper
fastapi~=0.115.12
uvicorn~=0.34.2
pydantic~=2.11.4